+++ execution/printing.c
@@ -1,11 +1,11 @@
 #include "../main.h"
 
-void print_current_elem(t_elem *ptr)
+void	print_current_elem(t_elem *ptr)
 {
 	dprintf(2,">>> %p, type %d next %p\n", ptr, ptr->type, ptr->next);
 }
 
-void print_elems(t_elem *root_ptr)
+void	print_elems(t_elem *root_ptr)
 {
 	dprintf(2,">>> \nprintelems\n");
 	t_elem *ptr= root_ptr;
